+++ gear/attribute_curator/src/python/attribute_curator_app/main.py
@@ -33,7 +33,8 @@ def run(context: GearToolkitContext,
     scheduler.apply(curator=curator, context=context)
 
     if curator.failed_files:
-        log.error(json.dumps(curator.failed_files, indent=4))
+        failed_files = curator.failed_files.copy()
+        log.error(json.dumps(failed_files, indent=4))
         raise GearExecutionError(
-            f"Failed to curate {len(curator.failed_files)} files, see above error logs"
+            f"Failed to curate {len(failed_files)} files, see above error logs"
         )